experiments/mcp-eval/pyproject.toml
[project]
name = "mcp-eval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
]
